requirement.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
bson
python-dotenv